+++ app-desktop/src/main/kotlin/org/archivekeep/app/desktop/ui/dialogs/sync/text-utils.kt
@@ -1,19 +1,28 @@
 package org.archivekeep.app.desktop.ui.dialogs.sync
 
 import org.archivekeep.files.operations.AdditiveRelocationsSyncStep
+import org.archivekeep.files.operations.CompareOperation
 import org.archivekeep.files.operations.NewFilesSyncStep
 import org.archivekeep.files.operations.PreparedSyncOperation
 import org.archivekeep.files.operations.RelocationsMoveApplySyncStep
 
 fun describePreparedSyncOperation(a: PreparedSyncOperation) =
-    if (a.isNoOp()) {
-        "NoOP"
+    if (a.steps.isEmpty()) {
+        "Nothing to do"
     } else {
         a.steps.joinToString("\n") {
             when (it) {
                 is AdditiveRelocationsSyncStep -> "Duplicate ${it.relocations.size} files."
 
-                is RelocationsMoveApplySyncStep -> "Move ${it.relocations.size} files."
+                is RelocationsMoveApplySyncStep -> {
+                    var text = "Move ${it.toApply.size} files."
+
+                    if (it.toIgnore.isNotEmpty()) {
+                        text += "Ignore ${it.toIgnore.size} relocations."
+                    }
+
+                    text
+                }
 
                 is NewFilesSyncStep -> "Upload ${it.unmatchedBaseExtras.size} files."
             }
@@ -23,8 +32,8 @@ fun describePreparedSyncOperation(a: PreparedSyncOperation) =
 fun describePreparedSyncOperationWithDetails(
     a: PreparedSyncOperation,
     action: String = "Add",
-) = if (a.isNoOp()) {
-    "NoOP"
+) = if (a.steps.isEmpty()) {
+    "Nothing to do"
 } else {
     a.steps
         .flatMap {
@@ -36,16 +45,44 @@ fun describePreparedSyncOperationWithDetails(
                         }
 
                 is RelocationsMoveApplySyncStep ->
-                    listOf("Move ${it.relocations.size} files.") +
-                        it.relocations.map { relocation ->
-                            " - move from ${relocation.extraOtherLocations} to ${relocation.extraBaseLocations}"
+                    (
+                        if (it.toApply.isNotEmpty()) {
+                            listOf("Execute ${it.toApply.size} relocations:")
+                        } else {
+                            emptyList()
                         }
+                    ) + (
+                        it.toApply.map { relocation ->
+                            "- ${relocation.describe()}"
+                        }
+                    ) + (
+                        if (it.toIgnore.isNotEmpty()) {
+                            listOf("Ignore ${it.toIgnore.size} relocations:")
+                        } else {
+                            emptyList()
+                        }
+                    ) + (
+                        it.toIgnore.map { relocation ->
+                            " - ignore ${relocation.describe()}"
+                        }
+                    )
 
                 is NewFilesSyncStep ->
-                    listOf("${action.capitalize()} ${it.unmatchedBaseExtras.size} files.") +
+                    listOf("${action.capitalize()} ${it.unmatchedBaseExtras.size} files:") +
                         it.unmatchedBaseExtras.map { group ->
                             " - $action new ${group.filenames}"
                         }
             }
         }.joinToString("\n")
 }
+
+fun CompareOperation.Result.Relocation.describe() =
+    (
+        if (isIncreasingDuplicates) {
+            "duplication increase from $otherFilenames to $baseFilenames"
+        } else if (isDecreasingDuplicates) {
+            "duplication decrease from $otherFilenames to $baseFilenames"
+        } else {
+            "move from $extraOtherLocations to $extraBaseLocations"
+        }
+    )