+++ files/src/main/kotlin/org/archivekeep/testing/storage/SpeedLimitedRepoWrapper.kt
@@ -41,8 +41,13 @@ class SpeedLimitedRepoWrapper(
         base.save(filename, info, stream)
     }
 
+    override suspend fun delete(filename: String) =
+        delayed(25) {
+            base.delete(filename)
+        }
+
     override suspend fun getMetadata(): RepositoryMetadata =
-        delayed(150) {
+        delayed(50) {
             base.getMetadata()
         }