+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ### Added
 
+- feature: relocations with duplication decrease support
+- UI: support relocations in Upload/Download dialogs (including duplication increase and/or decrease)
 - UI: Material Design TopAppBar (with client side decorations)
 - build: DEB and RPM packaging support
 - build: Maven Central publishing support